Fix error for tabgenie --help and minor refactoring
Refactoring towards faster flask cli

flask CLI could be really slow if
each command imports all the libraries also
from the other commands.

Solution: each command should import only its dependencies

This commit tried to refactored in this direction.
But it seems that we need create_app on most of the commands
For these commands it did not speed up the CLI
